% Generated by build script
\name{cdc_columns}
\alias{cdc_columns}
\title{Get column metadata for a CDC dataset}
\usage{
cdc_columns()
}
\arguments{
\item{dataset_id}{Socrata dataset ID (e.g. "vbim-akqf")}
}
\value{tibble: field_name, name, datatype, description}
\description{Get column metadata for a CDC dataset}
