% Generated by build script
\name{cdc_covid_cases}
\alias{cdc_covid_cases}
\title{COVID-19 case surveillance data}
\usage{
cdc_covid_cases()
}
\arguments{
\item{state}{State name filter (e.g. "California")}
\item{sex}{Sex filter: "Male", "Female"}
\item{age_group}{Age filter: "0 - 9 Years", "10 - 19 Years", etc.}
\item{where}{Additional SoQL WHERE clause}
\item{token}{Socrata app token}
\item{max_results}{Max results (default 1000)}
}
\value{tibble of COVID-19 case records}
\description{COVID-19 case surveillance data}
