% Generated by build script
\name{cdc_covid_vaccinations}
\alias{cdc_covid_vaccinations}
\title{COVID-19 vaccination data by jurisdiction}
\usage{
cdc_covid_vaccinations()
}
\arguments{
\item{location}{State/territory abbreviation (e.g. "CA", "NY")}
\item{token}{Socrata app token}
\item{max_results}{Max results (default 1000)}
}
\value{tibble of vaccination counts by date and jurisdiction}
\description{COVID-19 vaccination data by jurisdiction}
