% Generated by build script
\name{cdc_nndss}
\alias{cdc_nndss}
\title{NNDSS (National Notifiable Diseases Surveillance System) weekly data}
\usage{
cdc_nndss()
}
\arguments{
\item{disease}{Disease label filter (partial match)}
\item{year}{Year filter}
\item{token}{Socrata app token}
\item{max_results}{Max results (default 1000)}
}
\value{tibble of weekly disease counts by state}
\description{NNDSS (National Notifiable Diseases Surveillance System) weekly data}
